# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Music API",
    "version": "1.0.0",
    "description": "A RESTful API for a music streaming service, built with Node.js, Express, MongoDB and Cloudinary",
    "contact": {
      "name": "Music API Support"
    }
  },
  "servers": [
    {
      "url": "https://music-api-qxuc.vercel.app",
      "description": "Production server"
    },
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token"
      }
    },
    "schemas": {
...
      "Artist": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Artist ID"
          },
          "name": {
            "type": "string",
            "description": "Artist name"
          },
          "bio": {
            "type": "string",
            "description": "Artist biography"
          },
          "image": {
            "type": "string",
            "format": "uri",
            "description": "URL to artist image"
          },
          "genres": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of genre tags"
          },
          "followers": {
            "type": "number",
            "description": "Number of followers",
            "default": 0
          },
          "albums": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of album IDs"
          },
          "isVerified": {
            "type": "boolean",
            "description": "Whether artist is verified",
            "default": false
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "Song": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "description": "Song ID"
          },
          "title": {
            "type": "string",
            "description": "Song title"
          },
          "artist": {
            "type": "string",
            "description": "Artist ID"
          },
          "album": {
            "type": "string",
            "description": "Album ID"
          },
          "duration": {
            "type": "number",
            "description": "Duration in seconds"
          },
          "audioUrl": {
            "type": "string",
            "format": "uri",
            "description": "URL to audio file"
          },
          "coverImage": {
            "type": "string",
            "format": "uri",
            "description": "URL to cover image"
          },
          "releaseDate": {
            "type": "string",
            "format": "date-time",
            "description": "Release date"
          },
          "genre": {
            "type": "string",
            "description": "Song genre"
          },
          "plays": {
            "type": "number",
            "description": "Number of plays",
            "default": 0
          },
          "likes": {
            "type": "number",
            "description": "Number of likes",
            "default": 0
          },
          "isExplicit": {
            "type": "boolean",
            "description": "Whether song contains explicit content",
            "default": false
          },
          "featuredArtists": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Array of featured artist IDs"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message"
          },
          "status": {
            "type": "string",
            "description": "Error status"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/artists/{id}/top-songs": {
      "get": {
        "summary": "Get artist's top songs",
        "description": "Retrieve the most popular songs by a specific artist",
        "tags": [
          "Artists"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Artist ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Artist's top songs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Song"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Artist not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/artists/{id}": {
      "get": {
        "summary": "Get artist by ID",
        "description": "Retrieve detailed information about a specific artist",
        "tags": [
          "Artists"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Artist ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Artist retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artist"
                }
              }
            }
          },
          "404": {
            "description": "Artist not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update artist details",
        "description": "Update an existing artist's information",
        "tags": [
          "Artists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Artist ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "bio": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Artist image file"
                  },
                  "genres": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "isVerified": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Artist updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artist"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Artist not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an artist (Admin only)",
        "description": "Delete an artist from the database",
        "tags": [
          "Artists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Artist ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Artist deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Artist not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/artists/create": {
      "post": {
        "summary": "Create a new artist (Admin only)",
        "description": "Create a new artist profile with name, bio, and optional image",
        "tags": [
          "Artists"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "bio"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Artist name"
                  },
                  "bio": {
                    "type": "string",
                    "description": "Artist biography"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Artist image file"
                  },
                  "genres": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "isVerified": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Artist created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artist"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}